mudanca de int para float

+++ media_simples.c
@@ -1,5 +1,5 @@
-
-int media(int num1, int num2, int num3) {
+//Media Simples
+float media(float num1, float num2, float num3) {
     float num1, num2, num3, media;
 
     media = (num1 + num2 + num3) / 3;
